+++ photons_tile_paint/set_state.py
@@ -1,0 +1,229 @@
+"""
+Tile animations requires packing many SetState64 messages in a short amount of
+time. This file exists to make that as efficient as possible by creating a class
+that can pack a SetState64 for tile animation with as much caching power as
+possible
+
+Don't use this directly, always use it via the
+photons_tile_paint.animation.canvas_to_msgs function
+"""
+from photons_messages.fields import Color as ProtocolColor
+from photons_messages import TileMessages
+from photons_protocol.messages import T
+
+from input_algorithms import spec_base as sb
+from bitarray import bitarray
+from lru import LRU
+import functools
+import binascii
+import operator
+import struct
+
+class SetState64Maker:
+    """
+    One of these is made as the set_state_64_maker callable. It contains as much
+    of the SetState64 message packd as we can and then provides a fake SetState64
+    that photons can use to create the messages
+    """
+    class Target:
+        """Used to cache the bits for our target"""
+        def __init__(self, val):
+            self.bytes = val
+            self.serial = binascii.hexlify(val[:6]).decode()
+
+            self.packd = bitarray(endian="little")
+            self.packd.frombytes(val)
+
+    class SetState64:
+        """
+        The actual fake message
+
+        We assume this is used via canvas_to_msgs and so we expect kwargs to
+        contain "ack_required", "tile_index", "width", "duration" and "colors"
+
+        And we assume target is set via a setter property and .update() is
+        called with source and sequence
+
+        Because this is for tile animations, the rest of the payload is always
+        length=1, x=0, y=0; and res_required is always False
+
+        We use a real TileMessages.SetState64 to get the packed bits for the parts
+        that do not change, which is stored on the maker (the instance of SetState64Maker)
+        """
+        def __init__(self, maker, **kwargs):
+            if not hasattr(maker, "wat"):
+                maker.wat = True
+                self.wat = True
+            self.kwargs = kwargs
+            self.maker = maker
+            self.extra = {}
+
+        def pack(self):
+            bits = []
+
+            # Frame header
+            bits.append(self.maker.frame_header_start)
+            bits.append(self.maker.source_bits(self.extra["source"]))
+
+            # Frame Address
+            bits.append(self.extra["target"].packd)
+            if self.kwargs["ack_required"]:
+                bits.append(self.maker.frame_address_with_acks_middle)
+            else:
+                bits.append(self.maker.frame_address_without_acks_middle)
+            bits.append(self.maker.sequence_bits[self.extra["sequence"]])
+
+            # Protocol Header
+            bits.append(self.maker.protocol_header_packd)
+
+            # Payload
+            bits.append(self.maker.tile_index_bits[self.kwargs["tile_index"]])
+            bits.append(self.maker.payload_middle)
+            bits.append(self.maker.width_bits[self.kwargs["width"]])
+            bits.append(self.maker.duration_bits(self.kwargs["duration"]))
+            bits.append(self.maker.colors_bits(self.kwargs["colors"]))
+
+            return functools.reduce(operator.add, bits)
+
+        def tobytes(self, serial=None):
+            return self.pack().tobytes()
+
+        def simplify(self):
+            return self
+
+        def clone(self):
+            res = self.__class__(self.maker, **self.kwargs)
+            res.update(self.extra)
+            return res
+
+        def update(self, extra):
+            self.extra.update(extra)
+
+            if "target" in extra and not isinstance(extra["target"], self.maker.Target):
+                self.target = extra["target"]
+
+        @property
+        def target(self):
+            if "target" in self.extra:
+                return self.extra["target"].bytes
+            return sb.NotSpecified
+
+        @target.setter
+        def target(self, val):
+            self.update({"target": self.maker.make_target(val)})
+
+        @property
+        def serial(self):
+            if "target" not in self.extra:
+                return "000000000000"
+            return self.extra["target"].serial
+
+        @property
+        def res_required(self):
+            return False
+
+        @property
+        def ack_required(self):
+            return self.kwargs["ack_required"]
+
+        @property
+        def source(self):
+            return self.extra["source"]
+
+        @source.setter
+        def source(self, val):
+            """Source is set on the clone by the photons_transport writer"""
+            self.extra["source"] = val
+
+        @property
+        def sequence(self):
+            return self.extra["sequence"]
+
+        @property
+        def is_dynamic(self):
+            return False
+
+    def __init__(self):
+        self.targets_cache = LRU(1000)
+        self.source_bits_cache = LRU(10)
+        self.duration_bits_cache = LRU(10)
+
+        self.cache = ProtocolColor.Meta.cache
+
+        msg = TileMessages.SetState64(
+              source = 0
+            , sequence = 0
+            , target = None
+            , res_required = False
+            , ack_required = True
+
+            , tile_index = 0
+            , length = 1
+            , x = 0
+            , y = 0
+            , width = 8
+            , duration = 0
+            , colors = [{"hue": 0, "saturation": 0, "brightness": 0, "kelvin": 3500}]
+            )
+
+        self.frame_header_start = msg.frame_header.pack()[:-32]
+
+        self.frame_address_with_acks_middle = msg.frame_address.pack()[64:-8]
+        msg.ack_required = False
+        self.frame_address_without_acks_middle = msg.frame_address.pack()[64:-8]
+
+        self.protocol_header_packd = msg.protocol_header.pack()
+
+        # tile_index, width, duration and colors are variable
+        self.payload_middle = msg.payload.pack()[8:-8 -32 - (64 * 64)]
+
+        self.uint8_bits = {val: self.bits(T.Uint8, val) for val in range(256)}
+
+        self.width_bits = self.uint8_bits
+        self.sequence_bits = self.uint8_bits
+        self.tile_index_bits = self.uint8_bits
+
+    def bits(self, typ, val):
+        bits = bitarray(endian="little")
+        bits.frombytes(struct.pack(typ.struct_format, val))
+        return bits
+
+    def source_bits(self, val):
+        if val not in self.source_bits_cache:
+            self.source_bits_cache[val] = self.bits(T.Uint32, val)
+        return self.source_bits_cache[val]
+
+    def duration_bits(self, val):
+        if val not in self.duration_bits_cache:
+            self.duration_bits_cache[val] = self.bits(T.Uint32, int(val * 1000))
+        return self.duration_bits_cache[val]
+
+    def colors_bits(self, colors):
+        res = []
+        for color in colors:
+            dct = color.as_dict()
+            fields = tuple(sorted(dct.items()))
+            if fields not in self.cache:
+                self.cache[fields] = ProtocolColor(**dct).pack()
+            res.append(self.cache[fields])
+        return functools.reduce(operator.add, res)
+
+    def make_target(self, target):
+        if target not in self.targets_cache:
+            val = target
+            if isinstance(target, str):
+                val = binascii.unhexlify(target)
+
+            if isinstance(val, bytes) and len(val) == 6:
+                val += b"\x00\x00"
+
+            self.targets_cache[target] = self.Target(val)
+        return self.targets_cache[target]
+
+    def __call__(self, **kwargs):
+        return self.SetState64(self, **kwargs)
+
+# Used to create a message that photons thinks is a valid SetState64
+# It is a callable that requires keyword arguments for
+# ack_required, tile_index, width, duration, colors
+set_state_64_maker = SetState64Maker()